Rename instances of 'badge' in code to 'memento'
